% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fg_project.R
\name{num_summary}
\alias{num_summary}
\title{Summary of data}
\usage{
num_summary(x = "TSLA", start_date = "2011-01-03", end_date = Sys.Date())
}
\arguments{
\item{x}{string ticker symbol from yahoo finance}

\item{start_date}{string "YYYY-MM-DD" of starting date}

\item{end_date}{string "YYYY-MM-DD" of end date}
}
\value{
numerical summary of the data
}
\description{
Summary of data
}
\examples{
num_summary("^GSPC")
num_summary("AAPL", "2022-01-01", "2022-02-01")
}
