% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fg_project.R
\name{visualize_data}
\alias{visualize_data}
\title{Visualize data}
\usage{
visualize_data(x = "TSLA", start_date = "2011-01-03", end_date = Sys.Date())
}
\arguments{
\item{x}{string ticker symbol from yahoo finance}

\item{start_date}{string "YYYY-MM-DD" of starting date}

\item{end_date}{string "YYYY-MM-DD" of end date}
}
\value{
visualization of data: barplot, price chart and fear&greed index
}
\description{
Visualize data
}
\examples{
visualize_data("^GSPC")
visualize_data("AAPL", "2022-01-01", "2022-02-01")
}
