% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_sum.R
\name{weight_sum}
\alias{weight_sum}
\title{Title}
\usage{
weight_sum(x, weights = rep(1, length(x)), ...)
}
\arguments{
\item{x}{numeric vector}

\item{weights}{numeric vector of weight with length of x}

\item{...}{}
}
\value{
weighted sum
}
\description{
Title
}
\examples{
weight_sum(1:10)
}
